requirements.txt
pandas
requests
bs4
base64
streamlit
plotly.express

